karam-hackrx-project/requirements.txt
streamlit
python-dotenv
langchain
langchain-openai
langchain-community
langchain-core
langchain-chroma
langchain-text-splitters
openai
chromadb
pysqlite3-binary
pypdf
pandas
numpy
requests
pydantic
typing-extensions
aiohttp
httpx
